azure-functions-deploy/requirements.txt
azure-functions==1.19.0
azure-core>=1.30.0
